Added equatable conformance for tests

+++ CatchTheBus/Shared/Servers/TripsServer/Models/Trips/App/AppLocation.swift
@@ -18,7 +18,7 @@ struct AppLocation: Decodable, Hashable {
     let code: String
     let codeDetail: String
     let timezone: String
-    let heading: Int
+    let heading: Double
     let zone: [AppCoordinateModel]
     let hasFutureActivity: Bool
     let areaId: Int
